+++ modules/ml_services/ml_network/data.tf
@@ -1,3 +1,7 @@
 data "aws_subnet" "subnet_mlservices_1" {
   id = "${var.mlservices_subnet_id_1}"
 }
+
+data "aws_subnet" "subnet_mlservices_2" {
+  id = "${var.mlservices_subnet_id_2}"
+}