+++ modules/ml_services/ml_userdata/variables.tf
@@ -2,4 +2,6 @@ variable "api_port" {}
 variable "log_group" {}
 variable "proxy_endpoint" {}
 variable "service_name" {}
+variable "statsite_ip" {}
+variable "statsite_prefix" {}
 variable "tfs_port" {}